requirements.txt
django
requests 
django-rest-framework
django-cors-headers